requirements.txt
flask>=3.0
flask-cors>=4.0
google-genai>=1.0
chromadb>=1.5
requests>=2.31
beautifulsoup4>=4.12
